chore: create RecordAudioButton
